requirements.txt
selenium
webdriver-manager
pyyaml
openai  # Optional, for AI-powered evaluation if using OpenAI API